Allow the deletion of an image only to the user who owns it.

+++ app/data/operations.py
@@ -1,8 +1,9 @@
-
+import http
 import os
 import imghdr
 
 from fastapi import File, UploadFile, HTTPException
+from sqlalchemy.sql import select
 
 from app.config import IMAGES_FOLDER
 from app.data.models import UserInDB
@@ -28,7 +29,8 @@ async def add_image(file: UploadFile = File(...), user: UserInDB = None):
     # Check if uploaded file is actually an image
     img_type = imghdr.what(None, contents)
     if img_type is None:
-        raise HTTPException(status_code=422, detail="Uploaded file is not an image")
+        raise HTTPException(status_code=422,
+                            detail="Uploaded file is not an image")
     await save_file(relative_path, contents)
     query = images.insert().values(text="Image",
                                    relative_path=relative_path,
@@ -37,9 +39,18 @@ async def add_image(file: UploadFile = File(...), user: UserInDB = None):
     return last_record_id
 
 
-async def remove_image(id_: int, relative_path: str) -> None:
-    # Delete image file from file system
-    await delete_file(relative_path)
-    # Delete image record from images table
-    query = images.delete().where(images.columns.id == id_)
-    await database.execute(query)
+async def remove_image(id_: int, relative_path: str, user: UserInDB) -> None:
+    query = select([images.c.id, images.c.user_id]).where(images.c.id == id_)
+    db_images = await database.fetch_one(query)
+    # If a record is found
+    if db_images is not None:
+        # Do not perform the deletion of an image if the user who
+        # intends to perform the deletion does not own the image
+        if db_images['user_id'] != user.id:
+            raise HTTPException(status_code=http.HTTPStatus.UNAUTHORIZED,
+                                detail="Operation not allowed")
+        # Delete image file from file system
+        await delete_file(relative_path)
+        # Delete image record from images table
+        query = images.delete().where(images.columns.id == id_)
+        await database.execute(query)
